feat!: Made redux-offline compatible with expo SDK 36 by getting AsyncStorage back from react-native.
Expo SDK 36 upgraded to RN 0.60, but is not yet compatible with '@react-native-community/async-storage', so to make the repo compatible the deprecated AsyncStorage from 'react-native' has to be used.

This is a TEMPORAL SOLUTION until expo supports the new way of AsyncStorage.
Added folder lib to github to be able to access to it without publishing this fork to npm.